PhotoPortal/requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
pyodbc==4.0.30
azure-common==1.1.27
azure-core==1.12.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.8.0
Pillow==8.1.2
opencv-python==4.5.1.48
requests==2.25.1
requests-oauthlib==1.3.0